+++ src/test/java/racingcar/controller/RacingCarWebControllerTest.java
@@ -7,9 +7,9 @@ import org.junit.jupiter.api.Test;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
-import racingcar.controller.dto.CarStatusResponse;
-import racingcar.controller.dto.GameInfoRequest;
-import racingcar.controller.dto.RaceResultResponse;
+import racingcar.service.dto.CarStatusResponse;
+import racingcar.service.dto.GameInfoRequest;
+import racingcar.service.dto.RaceResultResponse;
 import racingcar.service.RaceResultService;
 
 import java.util.List;
